feat(common): make TextInput in common
